requirements.txt
Flask
Flask-Cors
Flask-Caching
requests
python-dotenv